feat: 영수증 등록 - 멤버 추가

+++ BackEnd/src/main/java/com/orange/fintech/payment/dto/ReceiptDetailMemberDto.java
@@ -26,5 +26,5 @@ public class ReceiptDetailMemberDto {
     String thumbnailImage;
 
     @Schema(description = "낼 금액")
-    int amountDue;
+    Long amountDue;
 }
