+++ BackEnd/src/main/java/com/orange/fintech/payment/entity/ReceiptDetailMember.java
@@ -14,5 +14,5 @@ public class ReceiptDetailMember {
 
     @EmbeddedId private ReceiptDetailMemberPK receiptDetailMemberPK;
 
-    private int amountDue;
+    private Long amountDue;
 }
